Make optimum_intervals a class method of Optimizer
This function is for now only ever used on Optimizer objects and accesses internals.
